Noncompiling, WIP! Use Maven and upgrade JavaParser
This code does *not* compile as is and contains some silly introduced errors.

This is a work-in-progress, switching to a supported version of JavaParser.

It also introduces the use of Maven as an alternative for building and
handling dependencies outside of Netbeans. One advantage of this approach
is to automatically take care of making needed dependencies available.